+++ src/pages/MePage/MePage.tsx
@@ -1,9 +1,9 @@
 import { Box, Button, Heading } from "grommet";
 import React, { useContext } from "react";
-import { MeBalances } from "../components/MeBalances";
-import { BrokContext } from "../context/BrokContext";
-import { SymfoniContext } from "../context/SymfoniContext";
-import { useSignAccessVP } from "../sign/useSignAccessVP";
+import { MeBalances } from "../../components/MeBalances";
+import { BrokContext } from "../../context/BrokContext";
+import { SymfoniContext } from "../../context/SymfoniContext";
+import { useSignAccessVP } from "./hooks/useSignAccessVP";
 var debug = require("debug")("page:me");
 
 interface Props {}
